Reverting move of helper.rb.
