+++ recipes/config-collectd.rb
@@ -7,7 +7,7 @@
 #
 # Copyright (c) 2015 SignalFx, Inc, All Rights Reserved.
 
-require File.expand_path("../../libraries/helper.rb", __FILE__)
+require File.expand_path("../helper.rb", __FILE__)
 
 def get_collectd_conf_template
   case node['platform_family']
