+++ recipes/config-vmstat.rb
@@ -7,7 +7,7 @@
 #
 # Copyright (c) 2016 SignalFx, Inc, All Rights Reserved.
 
-require File.expand_path("../../libraries/helper.rb", __FILE__)
+require File.expand_path("../helper.rb", __FILE__)
 
 include_recipe 'chef_install_configure_collectd::default'
 install_package_on_redhat 'epel-release'
